"duration2sample()": Edited.
Data type of return variable turned to <single> instead of <uint16>.

 Changes to be committed:
	modified:   duration2sample.m


Former-commit-id: a8268d114e2dd81a40be7acf14fb0fd490428d73

+++ lib/GA_parameterSearch/duration2sample.m
@@ -20,13 +20,13 @@
 function sampleCount = duration2sample(timeDuration, mode, SAMPLE_PERIOD)
 	% Sample count of appliances determined with 'ceil' method.
 	% However, time vector generated with 'floor' method.
-	sampleCount = uint16(ceil(minutes(timeDuration)/SAMPLE_PERIOD));
+	sampleCount = single(ceil(minutes(timeDuration)/SAMPLE_PERIOD));
 	
 	if strcmp(mode, '24h')
 		% If <sampleCount> equals to 0 (zero), then increae by one.
 		% MATLAB does not accept zero based indexing.
 		if sampleCount == 0
-			sampleCount = uint16(1);
+			sampleCount = single(1);
 		end
 	elseif ~strcmp(mode, 'inf')
 		error('duration2sample(): Undefined <mode>');
